requirements.txt
anyio==4.8.0
bcrypt==4.2.1
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.95.2
google-api-core==2.24.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.11.0
google-cloud-pubsub==2.17.0
google-cloud-storage==2.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
passlib==1.7.4
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==1.10.21
PyJWT==2.7.0
pyparsing==3.2.1
python-dotenv==1.0.1
python-multipart==0.0.5
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.22.0
pydantic[email]==1.10.21